Veliki deo security-a

+++ tim22/src/main/java/com/mrsisa/tim22/dto/UserTokenState.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class UserTokenState {
 
     private String accessToken;
-    private Long expiresIn;
+    private int expiresIn;
